+++ crates/factorio-mod-api/src/lib.rs
@@ -12,6 +12,7 @@ use std::{
 use api::{ApiToken, FullModSpec};
 use elsa::FrozenMap;
 use futures::StreamExt;
+use reqwest::Url;
 use semver::Version;
 use thiserror::Error;
 use tracing::info;
@@ -23,17 +24,32 @@ use crate::api::LoginResponse;
 pub struct ModPortalClient {
     client: reqwest::Client,
     specs: FrozenMap<String, Box<FullModSpec>>,
+    mod_api_base: Url,
+    auth_api_base: Url,
 }
 
 impl ModPortalClient {
     /// Creates a new client with default configuration.
     pub fn new() -> Result<ModPortalClient> {
-        ModPortalClient::with_client(reqwest::Client::builder().build()?)
+        ModPortalClient::with_client(reqwest::Client::default())
     }
 
     /// Creates a new client with a pre-configured `reqwest::Client`.
     pub fn with_client(client: reqwest::Client) -> Result<ModPortalClient> {
-        Ok(ModPortalClient { client, specs: FrozenMap::new() })
+        ModPortalClient::with_base_urls(
+            client,
+            Url::parse("https://mods.factorio.com").unwrap(),
+            Url::parse("https://auth.factorio.com").unwrap(),
+        )
+    }
+
+    /// Creates a new client, allowing substitution of custom base URLs.
+    fn with_base_urls(
+        client: reqwest::Client,
+        mod_api_base: Url,
+        auth_api_base: Url,
+    ) -> Result<ModPortalClient> {
+        Ok(ModPortalClient { client, specs: FrozenMap::new(), mod_api_base, auth_api_base })
     }
 
     /// Get the full spec of a Factorio mod. Request results are cached in memory.
@@ -56,7 +72,10 @@ impl ModPortalClient {
             spec
         } else {
             info!("requesting mod spec for '{name}'");
-            let url = format!("https://mods.factorio.com/api/mods/{name}/full");
+            let url = self
+                .mod_api_base
+                .join(&format!("api/mods/{name}/full"))
+                .map_err(|_| FactorioModApiError::InvalidModName { name: name.into() })?;
             let response = self.client.get(url).send().await?.json().await?;
             self.specs.insert(name.into(), Box::new(response))
         })
@@ -81,8 +100,7 @@ impl ModPortalClient {
     /// ```
     pub async fn login(&self, user_name: &str, password: &str) -> Result<ApiToken> {
         info!("logging in with user name '{user_name}'");
-
-        let url = "https://auth.factorio.com/api-login";
+        let url = self.auth_api_base.join("api-login").unwrap();
         let query = [("api_version", "4"), ("username", user_name), ("password", password)];
 
         let request = self.client.post(url).query(&query);
@@ -128,7 +146,10 @@ impl ModPortalClient {
             return Err(FactorioModApiError::InvalidModVersion { version: version.clone() })
         };
 
-        let url = format!("https://mods.factorio.com/{}", release.download_url);
+        let url = self
+            .mod_api_base
+            .join(&release.download_url)
+            .expect("the mod api shouldn't return invalid URLs");
         let query = [("username", &api_token.username), ("token", &api_token.token)];
 
         let response = self.client.get(url).query(&query).send().await?;
@@ -156,6 +177,10 @@ pub enum FactorioModApiError {
     #[error("Invalid mod dependency: '{dep}'")]
     InvalidModDependency { dep: String },
 
+    // Error that is raised if a mod name can't be used to call the API.
+    #[error("Invalid mod name: '{name}'")]
+    InvalidModName { name: String },
+
     // Error that is raised if a mod version doesn't exist.
     #[error("Invalid mod version: '{version}'")]
     InvalidModVersion { version: Version },
@@ -177,4 +202,118 @@ pub enum FactorioModApiError {
 
     #[error("Error while doing an IO operation")]
     IOError(#[from] std::io::Error),
+}
+
+#[cfg(test)]
+mod tests {
+    use std::error::Error;
+
+    use chrono::{DateTime, Utc};
+    use httpmock::prelude::*;
+    use ordered_float::NotNan;
+    use pretty_assertions::assert_eq;
+    use reqwest::Url;
+    use semver::Version;
+
+    use crate::{
+        api::{ModManifest, ModMetadata, ModRelease, ModSpec, ModTag},
+        ModPortalClient,
+    };
+
+    fn setup() -> Result<(MockServer, ModPortalClient), Box<dyn Error>> {
+        let server = MockServer::start();
+        let client = ModPortalClient::with_base_urls(
+            reqwest::Client::default(),
+            Url::parse(&server.base_url())?,
+            Url::parse(&server.base_url())?,
+        )?;
+
+        Ok((server, client))
+    }
+
+    fn mock_full(server: &MockServer) -> httpmock::Mock {
+        server.mock(|when, then| {
+            when.method(GET).path("/api/mods/mymod/full");
+            then.status(200)
+                .header("content-type", "application/json")
+                .body(include_bytes!("tests/full.json"));
+        })
+    }
+
+    #[tokio::test]
+    async fn full_request() -> Result<(), Box<dyn Error>> {
+        let (server, client) = setup()?;
+        let mock = mock_full(&server);
+
+        let spec = client.get_mod_spec("mymod").await?;
+
+        assert_eq!(
+            spec.short_spec,
+            ModSpec {
+                metadata: ModMetadata {
+                    name: "mymod".into(),
+                    owner: "someone".into(),
+                    summary: "SUMMARY".into(),
+                    title: "TITLE".into(),
+                    category: Some("general".into()),
+                    downloads_count: 42,
+                },
+                releases: vec![
+                    ModRelease {
+                        download_url: "/download/mymod/bde93f095d1b53ed019fca5e".into(),
+                        file_name: "mymod_0.0.1.zip".into(),
+                        info_json: ModManifest {
+                            factorio_version: "0.14".into(),
+                            dependencies: Some(vec!["base >= 0.13.0".try_into()?]),
+                        },
+                        released_at: DateTime::from_utc(
+                            DateTime::parse_from_rfc3339("2022-06-14T11:45:45.165000Z")?
+                                .naive_utc(),
+                            Utc
+                        ),
+                        version: Version::parse("0.0.1")?,
+                        sha1: "65b0435dbd4fb0ab0ceea61549641bf6f7dce9d2".into(),
+                    },
+                    ModRelease {
+                        download_url: "/download/mymod/7303634ba9642c5a321a757e".into(),
+                        file_name: "mymod_0.0.2.zip".into(),
+                        info_json: ModManifest {
+                            factorio_version: "0.14".into(),
+                            dependencies: Some(vec!["base >= 0.13.0".try_into()?]),
+                        },
+                        released_at: DateTime::from_utc(
+                            DateTime::parse_from_rfc3339("2022-09-24T08:53:02.970000Z")?
+                                .naive_utc(),
+                            Utc
+                        ),
+                        version: Version::parse("0.0.2")?,
+                        sha1: "a3499805018d6acaa29c1fbeaae763e6d8ae4279".into(),
+                    }
+                ],
+                description: Some("DESCRIPTION".into()),
+                github_path: Some("not/existing".into()),
+                tag: Some(ModTag { name: "general".into() }),
+                score: NotNan::new(88.15f64)?,
+                thumbnail: Some(
+                    "/assets/6eadeac2dade6347e87c0d24fd455feffa7069f0.thumb.png".into()
+                ),
+            }
+        );
+
+        mock.assert();
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn request_get_cached() -> Result<(), Box<dyn Error>> {
+        let (server, client) = setup()?;
+        let mock = mock_full(&server);
+
+        client.get_mod_spec("mymod").await?;
+        client.get_mod_spec("mymod").await?;
+
+        mock.assert_hits(1);
+        Ok(())
+    }
 }